+++ shared/src/commonMain/kotlin/net.wwwhackcom/di/Module.kt
@@ -4,7 +4,7 @@ import net.wwwhackcom.network.Api
 import net.wwwhackcom.network.ApiImpl
 import net.wwwhackcom.repository.AuthRepository
 import net.wwwhackcom.repository.AuthRepositoryImpl
-import net.wwwhackcom.viewmodel.ViewModel
+import net.wwwhackcom.experience.login.LoginViewModel
 import org.koin.dsl.module
 
 /**
@@ -21,7 +21,7 @@ val repositoryModule = module {
 }
 
 val viewModelModule = module {
-    single { ViewModel(get()) }
+    single { LoginViewModel(get()) }
 }
 
 val appModule = listOf(
